move trade/catch-error -> opendex/catch-error
